refactor: PlayerName 객체 생성
