+++ src/main/java/blackjack/domain/participant/PlayerName.java
@@ -1,0 +1,35 @@
+package blackjack.domain.participant;
+
+public class PlayerName {
+    private static final String ERROR_MESSAGE_WITH_SPACE = "이름에 공백이 포함됩니다.";
+    private static final String SPACE = " ";
+    private static final String EMPTY = "";
+
+    private final String name;
+
+    public PlayerName(String name) {
+        this.name = name;
+        validate(name);
+    }
+
+    private void validate(String name) {
+        validateEmpty(name);
+        validateSpace(name);
+    }
+
+    private void validateEmpty(String name) {
+        if(name.equals(EMPTY)) {
+            throw new IllegalArgumentException("올바르지 않은 이름입니다.");
+        }
+    }
+
+    private void validateSpace(String name) {
+        if (name.contains(SPACE)) {
+            throw new IllegalArgumentException(ERROR_MESSAGE_WITH_SPACE);
+        }
+    }
+
+    public String getName() {
+        return name;
+    }
+}
